Complete feature

+++ commands/rallyPlan/rallyPlanUpdater.ts
@@ -1,25 +1,35 @@
-import { IRally } from "../../entities/Rally/Rally";
+import { TextChannel } from "discord.js";
+import { client } from "../../bot";
+import { Rally } from "../../entities/Rally/Rally";
 import { RallyPlan } from "../../entities/RallyPlan/RallyPlan";
+import { dmRallyReadyToUsers, generateRallyMessage } from "../rallyCommand/rallyCommandHelper";
 
-setTimeout(() => {
-
-}, 10000);
-
-const updateRally = async () => {
+const rallyPlanUpdater = async () => {
   const rallyPlans = await RallyPlan.find();
-  const now = Date.now();
 
   for (const rallyPlan of rallyPlans) {
-    if (rallyPlan.scheduledEpoch <= now) {
-      rallyPlan.deleteOne();
-    }
+    Rally.findOne({ _id: rallyPlan.rallyId }).then(async (rally) => {
+
+      const channel = await client.channels.fetch(rally.channelId) as TextChannel;
+
+      const message = await channel.messages.fetch(rally.messageId)
+
+      message.edit(generateRallyMessage(rally, rallyPlan));
+
+      if (rallyPlan && rallyPlan.scheduledEpoch <= Date.now()) {
+        if (rally.hasFilled) {
+          dmRallyReadyToUsers(rally)
+        }
+
+        console.info(`deleting ${rallyPlan._id}`);
+        rallyPlan.deleteOne();
+      }
+    });
   }
-}
 
-const updateRallyMessage = (rally: IRally, mins) => {
+  setTimeout(() => {
+    rallyPlanUpdater()
+  }, 15000);
+};
 
-}
-
-const updateRallyWithDoneMessage = (rally: IRally) => {
-
-}
+export default rallyPlanUpdater;
